refactor: expose transactions; allow unsigned transaction encoding with parameter
